A function to reverse a string
Let it never be said I couldn't reverse a string. This version uses a
bitwise exclusive or to perform an inplace swap of characters referenced
by the string (one front and one end pointer). The function was
originally written on an android phone during a very bumpy bus ride from
antalya to alanya.

Signed-off-by: Adam McCartney <adam@mur.at>