Beargit commit
In the spirit of the original this version is _really_ dumb. There
is lots of duplication of code and it is written to be as stupid as
possible. It does however do what it's supposed to do. See the readme on
details of the operations that should be completed once a beargit commit
has been successful.

Signed-off-by: Adam McCartney <adam@mur.at>