beargit: commit (cleanup)
Slight refactor of beargit commit in order to narrow down the search
area while looking for a bug related to the test suite. (Mysterious
SIGKILL was getting thrown because of a pthread error. Seems likey to
have had to do with not releasing the file lock for the index when doing
multiple commits in quick succession.)

Signed-off-by: Adam McCartney <adam@mur.at>